src/requirements.txt
matplotlib==3.3.4
numpy==1.19.2
opencv_python==4.5.4.60
pandas==1.1.5
Pillow==9.4.0
PyYAML==6.0
scikit_learn==1.2.0
seaborn==0.11.2
tensorboardX==2.5.1
torch==1.10.1
torchsummary==1.5.1
torchvision==0.11.2
